#2 refactor-feature/JD: int -> Long으로 수정

+++ src/main/java/com/zb/jogakjogak/resume/domain/ResumeResponseDto.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ResumeResponseDto {
-    private int resumeId;
+    private Long resumeId;
     private String name;
     private String content;
 }
